requirements.txt
flask
ufal.morphodita
